move conda_signal_handler to the base __init__.py
